% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_insulin_for_endogenous_cpeptide.R
\name{adjust_insulin}
\alias{adjust_insulin}
\title{Adjust for endogenous insulin production.}
\usage{
adjust_insulin(df, idv = "TAD", dosing_time = 0.01, dv = "CONC",
  dv_type = "TYPE", dv_type_values = list(INSULIN = "INSULIN", CPEPTIDE
  = "CPEPTIDE"))
}
\arguments{
\item{df}{a dataset for one subject with insulin and cpeptide values provided
in the same variable e.g. CONC}

\item{idv}{the time variable}

\item{dosing_time}{time in the idv variable at which insulin injection was
administered}

\item{dv}{the variable containing insulin and c-peptide values}

\item{dv_type}{a flag variable identifying insulin and c-peptide values}

\item{dv_type_values}{list of expression indicating the values of the dv_type
variable e.g. if values of dv_type are `ins` and `cp` then dv_type_values
will take the expression list("INSULIN"="ins", "CPEPTIDE"="cp")}
}
\value{
a dataframe with new variables INSCP_RAT
  (\eqn{\frac{insulin}{cpeptide}}) and INSULIN_ADJ (adjusted insulin
  concentration)
}
\description{
This function calculates baseline insulin to C-Peptide ratio and adjust the
observed insulin for c-peptide (endogenous insulin) production
}
\examples{
# Test the function using the \\emph{wosmpkdata} dataframe.
pt1<-wosmpkdata\%>\%filter(ID==1, PERIOD==1)
pt1<-adjust_insulin(pt1)
}
