% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smooth_mean_plot_and_tableone.R
\name{base_mean_plot}
\alias{base_mean_plot}
\title{Plot mean concentration vs time profile.}
\usage{
base_mean_plot(df, treatment = "TREATMENT", time = "TAD",
  dv = "CONC")
}
\arguments{
\item{df}{a dataset with data for plotting mean profiles}

\item{treatment}{a variable name (string) for treatment groups}

\item{time}{a variable name for time after dose}

\item{dv}{a variable name for concentration column}
}
\value{
a plot
}
\description{
This function is a wrapper for ggplot and stat_summary to plot mean+-95%CI
concentration-vs-time profiles.
}
\examples{
base_mean_plot(wosmpkdata\%>\%filter(INCLUDED=="YES"))+
 #base_theme()+plot_theme()+
 labs(x="Time (hours)", y = "Concentrations (ng/mL)")+
 scale_color_manual(name=NULL, values = c("Mixtard 30/70" = "red", "Consegna 30/70" = "blue"))+
 facet_grid(PERIOD~TYPE, labeller = labeller(PERIOD=c("1"="Period 1", "2"= "Period 2")))
}
