% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEQR.R
\docType{data}
\name{beqrdemographic}
\alias{beqrdemographic}
\title{demographics of health volunteers in the PKPD study}
\format{a data-frame with 97 rows and 7 variables
\describe{
  \item{ID}{Unique subject identifier (1 -- 50)}
  \item{GENDER}{Biological sex, M = Male, F=Female}
  \item{AGE}{Age in years}
  \item{WEIGHT}{Weight in Kg}
  \iter{PERIOD}{Treatment periods}
  \item{SEQUENCE}{Assigned treatment sequences}
  \item{TREATMENT}{Treatments taken in each treatment period}
}}
\usage{
beqrdemographic
}
\description{
A dataset of demographic characteristics
}
\keyword{datasets}
