% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEQR.R
\docType{data}
\name{beqrpddataset}
\alias{beqrpddataset}
\title{Pharmacodynamic dataset from the PKPD study}
\format{a data-frame with 21160 rows and 8 variables
\describe{
  \item{ID}{Unique subject identifier (1 -- 50)}
  \item{PERIOD} {Treatment periods}
  \item{TIME}{Time of the day in 24 hours format}
  \item{GIR}{Glucose infusion rate, milliter/hour}
  \item{GLUCOSE}{Plasma glucose concentration, milgram/deciliter}
  \item{SEQUENCE}{Assigned treatment sequences}
  \item{TREATMENT}{Treatments taken in each treatment period}
  \item{CLOCK}{Time of the day in 24 hours format}
}}
\usage{
beqrpddataset
}
\description{
A dataset providing glucose infusion rate and plasma glucose
levels overtime during the observation period.
}
\keyword{datasets}
