% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEQR.R
\docType{data}
\name{beqrpkdataset}
\alias{beqrpkdataset}
\title{Pharmacokinetic dataset from the PKPD study}
\format{a data-frame with 4656 rows and 9 variables
\describe{
  \item{ID}{Unique subject identifier (1 -- 50)}
  \item{PERIOD}{Treatment periods}
  \item{TIME}{Time after insulin injection}
  \item{CONC}{Insulin or c-peptide concentration}
  \item{TAD}{Time after insulin injection, hours}
  \item{TYPE}{Analyte type, INSULIN or CPEPTIDE}
  \item{INCLUDED}{Observations included for analysis}
  \item{SEQUENCE}{Assigned treatment sequences}
  \item{TREATMENT}{Treatments taken in each treatment period}
}}
\usage{
beqrpkdataset
}
\description{
A dataset providing insulin and c-peptide concentrations
over-time during the observation period
}
\keyword{datasets}
