% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_model_explr_funtions.R
\name{custom_combo}
\alias{custom_combo}
\title{Plot continous versus categorical variables}
\usage{
custom_combo(data, mapping, ...)
}
\arguments{
\item{data}{dataset containg continous and categorical variables for which to
explore relationships}

\item{mapping}{aes mappings as in ggplot aes}
}
\value{
a ggplot2 object to plot
}
\description{
The function was written for exploration of parameter versus categorical
covariates relationships. It can be used to customize plots produced by the
GGally package
}
