% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_data_prep_funtions.R
\name{custom_round}
\alias{custom_round}
\title{Round time after dose to a specific base.}
\usage{
custom_round(x, base, cutoff)
}
\arguments{
\item{x}{a vector to round}

\item{base}{a base to round by}

\item{cutoff}{a cutoff (usually 0.5*base) to round by}
}
\value{
a vector with values rounded
}
\description{
This function rounds a value to a base value based on a cut-off value.
}
\examples{
custom_round(18, 5, 2.5)
custom_round(20, 5, 2.5)
}
