% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_data_prep_funtions.R
\name{edit_Date}
\alias{edit_Date}
\title{Edit date values after midnight}
\usage{
edit_Date(df, condition_col = "TIME", date_col = "DATE")
}
\arguments{
\item{df}{a dataframe with date column to edit}

\item{condition_col}{a column which provides a condition by which date-col
should be edited}

\item{date_col}{a date column to edit}
}
\value{
input dataframe df with date_col edited
}
\description{
Edit date values after midnight
}
