% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dose_partialAuc_and_extract_NCA_results.R
\name{extract_nca_results}
\alias{extract_nca_results}
\title{Extract and organize results from NCA analysis.}
\usage{
extract_nca_results(pknca_results, select_nca_param, design_var = c("ID",
  "TREATMENT", "PERIOD", "SEQUENCE"))
}
\arguments{
\item{pknca_results}{results from the function PKNCA::pk.nca functions}

\item{select_nca_param}{a character vector of PK parameters to be extracted
from PKNCA::pk.nca results}

\item{design_var}{a character vector of column names for BEQ study design
variables i.e ID, TREATMENT, PERIOD, and SEQUENCE}
}
\value{
a dataframe with NCA parameter values in a wide format
}
\description{
This function is for extracting results from NCA analysis and put them in
appropriate format for bioequivalence analysis.
}
