% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_bioequivalence_by_nlme.R
\name{get_abe_bylme}
\alias{get_abe_bylme}
\title{Get average bioequivalence results.}
\usage{
get_abe_bylme(data, treatments, sequences, periods, idvar, abe_param,
  reference)
}
\arguments{
\item{data}{a dataset with pharmacokinetic or pharmacodynamic parameters
whose ABE are to be calculated}

\item{treatments}{unquoted variable name for treatment groups (Should be a factor variable)}

\item{sequences}{unquoted variable name for treatment sequences(Should be a factor variable)}

\item{periods}{unquoted variable name for treatment periods (Should be a factor variable)}

\item{idvar}{unquoted variable name for subject identification}

\item{abe_param}{unquoted variable name for a PK/PD parameter whose
bioequivalence is sought}

\item{reference}{a value of the treatments variable to use as reference
treatment}
}
\value{
a list with 2 elements: lmefit = results from nlme::lme and ABEresult
  = a dataframe of ABE results
}
\description{
This function uses nlme::lme function to fit the linear mixed effects model.
The average bioequivalence results are calculated from the estimate of the
treatment effect and the corresponding 90% confidence interval
}
\examples{
fit<-get_abe_bylme(data=results_wide, treatments = TREATMENT, sequences = SEQUENCE, periods = PERIOD, idvar = ID, abe_param = cmax0_24, contr_variable = "Mixtard 30/70")
abe_res<-fit$ABEresult
modelfit<-fit$lmefit
}
