% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observed_predicted_indv_profile.R
\name{get_individual_plots}
\alias{get_individual_plots}
\title{Get individual plots from a dataset of multiple subjects.}
\usage{
get_individual_plots(df, id = id, .plotfun = plot_dv_ipred_pred,
  study = study, dose = dose, dv = dv, idv = idv, ylab = ylab,
  xlab = xlab)
}
\arguments{
\item{df}{data for a single patient per visit}

\item{id}{a variable name (string) for subject unique identifier}

\item{.plotfun}{a function to plot individual plots}

\item{study}{a study identification ID}

\item{dose}{a variable name (string) for dose variable}

\item{dv}{a variable name (string) for dependent variable}

\item{idv}{a variable name (string) for independet (regressor) variable}

\item{ylab}{a label for y-axis}

\item{xlab}{a label for x-axis}
}
\value{
a list of individual plots
}
\description{
Get individual plots from a dataset of multiple subjects.
}
