% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_individual_and_multiple_plots.R
\name{get_individual_profiles}
\alias{get_individual_profiles}
\title{Plot insulin profiles for all individuals and respective periods.}
\usage{
get_individual_profiles(df, id = "ID", visit = "PERIOD",
  form = "TREATMENT", dv = list(CONC = "CONC"), idv = "TIME",
  ylab = "Concentration (ng/mL)", xlab = "Time (hrs)", nplots = 4,
  ncol = 2, response_to_plot = "INSULIN", plot_function,
  adj_insert_fig_height)
}
\arguments{
\item{df}{dataset containing individuals with analysable data. Variable names
should be as given in the function.}

\item{id}{a variable name (string) for subject unique identifier}

\item{visit}{a variable name for treatment period}

\item{form}{a variable name (string) for formulations variable}

\item{dv}{a variable name (string) for dependent variable}

\item{idv}{a variable name (string) for independet (regressor) variable}

\item{ylab}{a label for y-axis}

\item{xlab}{a label for x-axis}

\item{nplots}{number of plots per panel}

\item{ncol}{number of plot-columns per panel}

\item{response_to_plot}{a response variable to be plotted (Insulin or Glucose
infusion rate)}

\item{plot_function}{a function used to plot individual profiles
(plot_insulin or plot_infusionrate)}
}
\value{
plots of individual profiles
}
\description{
This function uses the multiplot function and functions for plotting
individual profiles to create multiple plots for individuals.
}
\examples{
get_individual_profiles(insulinpk\%>\%filter(ID\%in\%c(1,2,3)), id="ID", form = "PERIOD", dv = list("CONC"="CONC"), idv = "TAD", period = "PERIOD",
                 nplots = 6, ncol = 2, ylab = "Concentration (ng/mL)", xlab="Time (hrs)", response_to_plot = "INSULIN", plot_function = plot_insulin)
get_individual_profiles(pddata\%>\%filter(ID\%in\%c(1,2,3)), id="ID", form = "TREATMENT", dv = list("GIR"="GIR", "GLUCOSE"="GLUCOSE", "CLAMP" = "GCLAMP"),
                 idv = "TAD", period = "PERIOD", nplots = 6, ncol = 2, ylab = "Glucose infusion rate (mL/h)", xlab="Time (hrs)",
                 plot_function = plot_infusionrate, adj_insert_fig_height = 20, response_to_plot = "GIR")
}
