% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smooth_mean_plot_and_tableone.R
\name{get_smooth_dv}
\alias{get_smooth_dv}
\title{Smooth observations using the LOESS procedure.}
\usage{
get_smooth_dv(df, idv, dv, ...)
}
\arguments{
\item{df}{a dataset with independent and dependent variable}

\item{idv}{indpendent variable (string)}

\item{dv}{dependent variable (string)}

\item{...}{other arguments to be passed to function stats::loess}
}
\value{
returns a the original dataframe containing a new variable of
  predicted dv (smooth_dv)
}
\description{
This is wrapper for stats::loess to create a loess model that is then used to
predict observations where they are not measured
}
\examples{
wosmpddata<-read.csv("../data/secdata/wosm105_gir_pddata.csv", stringsAsFactors = FALSE)
id1pd<-wosmpddata\%>\%filter(ID==1, PERIOD==1)
id1pd<-id1pd\%>\%group_by(ID,PERIOD)\%>\%do(get_smooth_dv(.,idv = "TAD", dv= "GIR", span = 0.1))
ggplot(aes(x=TAD, y= GIR), data = id1pd)+
geom_point()+
geom_line(aes(y=smooth_dv))
}
