% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_model_explr_funtions.R
\name{get_your_ggally_plots}
\alias{get_your_ggally_plots}
\title{Extract plots created by GGally package.}
\usage{
get_your_ggally_plots(plots, ncolx, ncoly)
}
\arguments{
\item{plots}{A GGally object containing ggplot2 plot objects}

\item{ncolx}{number of variables specified as x variables when creating a plot using ggally}

\item{ncoly}{number of variables specified as y variables when creating a plot using ggally}
}
\value{
a list of plots from the GGally object
}
\description{
Extract plots created by GGally package.
}
