% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dose_partialAuc_and_extract_NCA_results.R
\name{make_nca_dose_dataset}
\alias{make_nca_dose_dataset}
\title{Make dose dataset for NCA analysis with PKNCA.}
\usage{
make_nca_dose_dataset(df, treatment_var, id_var, time_var, dose, dose_time,
  tau)
}
\arguments{
\item{df}{a dataset}

\item{treatment_var}{a variable (unquoted name) with treatment groups}

\item{id_var}{a variable () with subjects identification numbers}

\item{time_var}{a variable () with time values}

\item{dose}{a dose value}

\item{dose_time}{a time of dose value}

\item{tau}{a value for interdose interval}
}
\value{
a dataframe with dosing information
}
\description{
This function makes a dosing dataset for use in NCA analysis by R PKNCA
package.
}
\examples{
dose<-make_nca_dose_dataset(insulinpk, treatment_var = TREATMENT, id_var = ID, time_var = TIME, dose = 7287700, dose_time = 0, tau = 12)
}
