% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dose_partialAuc_and_extract_NCA_results.R
\name{make_nca_par_dataset}
\alias{make_nca_par_dataset}
\title{Create partial AUC interval object.}
\usage{
make_nca_par_dataset(nca_params, partial_aucs = list(start, end),
  protocol_tlast, compute_aucinf = TRUE)
}
\arguments{
\item{nca_params}{PK parameters to be calculated by NCA analysis}

\item{partial_aucs}{a list object with start (start time to partial AUC) and
end (end time for partial AUC) elements}

\item{protocol_tlast}{a value of end of sampling time based on the protocol}

\item{compute_aucinf}{whether or not to compute AUCinf (TRUE or FALSE)}
}
\value{
a dataframe with NCA parameters to be computed by PKNCA and partial
  areas
}
\description{
A dataset with intended partial areas is required for NCA analysis using the
PKNCA package. This function creates such partial area dataset for PKNCA
package.
}
\examples{
#Explore NCA parameters to select names appropriately
nca_params<-PKNCA::get.interval.cols()
names(nca_params)
#Create the dataset with correct NCA parameter names
testdf<-make_nca_par_dataset(nca_params = c("cmax","tmax","aumclast", "lambda.z","half.life","aucpext.obs"),
                            partial_aucs = list(start=c(0,0,0), end=c(4, 6, 12)),
                            protocol_tlast = 24,
                            compute_aucinf = FALSE)
}
