% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_template_report.R
\name{make_template_report}
\alias{make_template_report}
\title{Create an example rmarkdown file report.}
\usage{
make_template_report(dir_path = "lab-notebook",
  file_name = "newfile.rmd")
}
\arguments{
\item{dir_path}{relative or absolute path to an existing file directory}

\item{file_name}{a name to be given to the template rmarkdown file (include
*.rmd extension)}
}
\value{
creates an rmarkdown file in an existing file directory
}
\description{
This function has a template rmarkdown file. The file contains BEQR functions
for plotting, NCA and ABE analyis.
}
