% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_bioequivalence_by_nlme.R
\name{make_treatment_contrast}
\alias{make_treatment_contrast}
\title{Make contrast matrix.}
\usage{
make_treatment_contrast(df, fact_var, contr_variable)
}
\arguments{
\item{df}{a dataset with factor variable to create contrast from}

\item{fact_var}{an unquote factor variable to create constrast}

\item{contr_variable}{a value of fact_var to use as a reference contrast}
}
\value{
a contrast matrix
}
\description{
This function enables the user to set a reference treatment for
bioequivalence comparison using get_abe_bylme to set reference treatment.
}
\examples{
make_treatment_contrast(df=results_wide, fact_var = TREATMENT, contr_variable = "Mixtard 30/70")
}
