% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_individual_and_multiple_plots.R
\name{multiplot}
\alias{multiplot}
\title{Plot mutiple plots in one panel.}
\usage{
multiplot(..., plotlist = NULL, file, cols = cols, layout = NULL)
}
\arguments{
\item{...:}{individula ggplot objects}

\item{plotlist:}{a list of ggplot objects}

\item{cols:}{Number of columns in layout}

\item{layout:}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
This function was accessed on November 2015 at
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_%28ggplot2%29/.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE), then
plot 1 will go in the upper left, 2 will go in the upper right, and 3 will go
all the way across the bottom.
}
