% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observed_predicted_indv_profile.R
\name{plot_dv_ipred_pred}
\alias{plot_dv_ipred_pred}
\title{Plot individual profiles.}
\usage{
plot_dv_ipred_pred(df, id = "ID", study = "STUDY", dose = "DOSE",
  dv = list(OBS = "DV", PRED = "PRED", IPRED = "IPRED"), idv = "TIME",
  ylab = "Concentration (ng/mL)", xlab = "Time (hrs)")
}
\arguments{
\item{df}{data for a single patient per visit}

\item{id}{a variable name (string) for subject unique identifier}

\item{study}{a study identification ID}

\item{dose}{a variable name (string) for dose variable}

\item{dv}{a variable name (string) for dependent variable}

\item{idv}{a variable name (string) for independet (regressor) variable}

\item{ylab}{a label for y-axis}

\item{xlab}{a label for x-axis}
}
\value{
a ggplot object
}
\description{
Plot individual profiles.
}
