% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_individual_and_multiple_plots.R
\name{plot_infusionrate}
\alias{plot_infusionrate}
\title{Plot individual subject's glucose infusion rate profile by treatment period.}
\usage{
plot_infusionrate(df, id = "ID", idv = "TAD", dv = list(GIR = "GIR",
  GLUCOSE = "GLUCOSE", CLAMP = "GCLAMP"), visit = "PERIOD",
  form = "TREATMENT", ylab = "Glucose infusion rate (mL/h)",
  xlab = "Time (hrs)", adj_insert_fig_height)
}
\arguments{
\item{df}{data for a single patient visit. e.g subject 1 period 1. Variable
names should be as given in the function.}

\item{id}{a variable name (string) for subject unique identifier}

\item{idv}{a variable name (string) for independet (regressor) variable}

\item{dv}{a variable name (string) for dependent variable}

\item{visit}{a variable name for treatment period}

\item{form}{a variable name (string) for formulations variable}

\item{ylab}{a label for y-axis}

\item{xlab}{a label for x-axis}

\item{adj_insert_fig_height}{adjustment for the height of the insert
glucose-profile figure.}
}
\value{
a ggplot object
}
\description{
Plot individual subject's glucose infusion rate profile by treatment period.
}
\examples{
pt1<-pddata\%>\%filter(ID==1&PERIOD==1)
plot_infusionrate(pt1, adj_insert_fig_height = 20)
}
