% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_individual_and_multiple_plots.R
\name{plot_insulin}
\alias{plot_insulin}
\title{Plot individual subject's insulin profile by treatment period.}
\usage{
plot_insulin(df, id = "ID", visit = "PERIOD", form = "TREATMENT",
  dv = list(CONC = "CONC"), idv = "TIME",
  ylab = "Concentration (ng/mL)", xlab = "Time (hrs)")
}
\arguments{
\item{df}{data for a single patient visit. e.g subject 1 period 1. Variable
names should be as given in the function.}

\item{id}{a variable name (string) for subject unique identifier}

\item{visit}{a variable name for treatment period}

\item{form}{a variable name (string) for formulations variable}

\item{dv}{a variable name (string) for dependent variable}

\item{idv}{a variable name (string) for independet (regressor) variable}

\item{ylab}{a label for y-axis}

\item{xlab}{a label for x-axis}
}
\value{
a ggplot2 object
}
\description{
Plot individual subject's insulin profile by treatment period.
}
\examples{
pt1<-insulinpk\%>\%filter(ID==1&PERIOD==1)
plot_insulin(pt1, visit = "PERIOD",form = "TREATMENT", idv = "TAD")
}
