% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smooth_mean_plot_and_tableone.R
\name{se}
\alias{se}
\title{Calculate standard errors of a vector}
\usage{
se(x)
}
\arguments{
\item{x}{a vector whose standard error is required}
}
\value{
standard error for the vector x
}
\description{
Calculate standard errors of a vector
}
