% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_data_prep_funtions.R
\name{str_panctuate}
\alias{str_panctuate}
\title{Insert panctuations in a string}
\usage{
str_panctuate(string, panc_pos, panctuation)
}
\arguments{
\item{string}{a string to be panctuated}

\item{panc_pos}{a position to insert panctuation}

\item{panctuation}{a panctuation to insert}
}
\value{
a panctuated string
}
\description{
This function can be used to panctuate dates into a format that can be
modified with lubridate
}
