% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nca_variable_statistics.R
\name{summarize_statz}
\alias{summarize_statz}
\title{Summarize statistics of continous variables contained in a dataframe.}
\usage{
summarize_statz(df, group, variables, stats = list("mean", "sd", "CV",
  "median", "min", "max", "geometric.mean"))
}
\arguments{
\item{df}{a dataframe with variables to perform descriptive stats}

\item{group}{a unquoted name of a grouping variable}

\item{variables}{a character vector of variables to be summarized.}

\item{stats}{descriptive summary statistics to compute. (NOT YET IMPLEMENTED. USED DEFAULTS ARE: "mean", "sd", "CV", "median", "min", "max", "quantiles (0.25 and 0.75)", "geometric.mean" and "Goemetric.cv")}
}
\value{
a dataframe of summary statistics
}
\description{
Summarize statistics of continous variables contained in a dataframe.
}
