% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smooth_mean_plot_and_tableone.R
\name{table_characteristics}
\alias{table_characteristics}
\title{Table subjects characteristcs.}
\usage{
table_characteristics(data, vars, factorVars, strata, ...)
}
\arguments{
\item{data}{a dataset with subjects characterics variables}

\item{vars}{a character vector of subject characteristics}

\item{factorVars}{a character vector of categorical variables}

\item{strata}{a character vector grouping variables}

\item{...}{other arguments passed to function tableone::CreateTableOne}
}
\value{
a dataframe of subject characteristics with descriptive and
  comparison statistics
}
\description{
This function is a wrapper for tableone::CreateTableOne and creates a table
of subject characteristics by treatment groups and/or sequence groups
}
\examples{
tab_char<-table_characteristics(vars = c("GENDER", "AGE", "WEIGHT"), strata = c("SEQUENCE", "PERIOD"), includeNA = TRUE, data = demographics)
}
