% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nca_variable_statistics.R
\name{table_median_tmax}
\alias{table_median_tmax}
\title{Table of Tmax descriptive stats}
\usage{
table_median_tmax(df, group, tmax = "tmax0_24")
}
\arguments{
\item{df}{a result from PKNCA::pk.nca with tmax and half-life parameters}

\item{group}{a column name (string) for treatment groups}

\item{tmax}{a column name (string) for tmax variable}
}
\value{
a dataframe of tmax and t-half or an error message
}
\description{
Table of Tmax descriptive stats
}
