% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nca_variable_statistics.R
\name{table_median_tmax_thalf}
\alias{table_median_tmax_thalf}
\title{Get median tmax and half-life}
\usage{
table_median_tmax_thalf(df, group, tmax = "tmax0_24",
  thalf = "half.life0_24")
}
\arguments{
\item{df}{a result from PKNCA::pk.nca with tmax and half-life parameters}

\item{group}{a column name (string) for treatment groups}

\item{tmax}{a column name (string) for tmax variable}

\item{thalf}{a column name (string) for half-life variable}
}
\value{
a dataframe of tmax and t-half or an error message
}
\description{
Get median tmax and half-life
}
\examples{
results_wide_pairs<-results_wide\%>\%group_by(ID)\%>\%mutate(N=n())\%>\%ungroup()\%>\%filter(N==2)\%>\%select(-N)
test<-table_median_tmax_thalf(results_wide_pairs, group = "TREATMENT")
}
